+++ src/swgoh_comlink/StatCalc/__init__.py
@@ -1,4 +1,5 @@
 from .calculator import StatCalc
 from .calculator_async import StatCalcAsync
+from .data_builder import GameDataBuilder, GameDataBuilderAsync
 
-__all__ = ["StatCalc", "StatCalcAsync"]
+__all__ = ["GameDataBuilder", "GameDataBuilderAsync", "StatCalc", "StatCalcAsync"]